Most recent version of SST
This update added:

Using delta time blocks in turbowarp to be as accurate as I can (For now).
After about a minute, the project is like 4 seconds off lol. This is assuming that scratches timer is accurate. I will test this later with like 5 different stopwatches to see how far off it ACTUALLY is, because scratches wait blocks are pretty bad so the timer is probably worse. This is why I changed the timing to broadcast whenever the number that came after it (Reading left to right) is equal to 0 which means that it's looped over the full amount of time...

None of that actually makes sense but hopefully you can work it out lol

Open the .sb3 in https://turbowarp.org/