FINAL VERSION FOR NOW
Contains screenshots of what I did

Open the .sb3 at https://turbowarp.org/

Stuff I added:

DARK MODE (I just got rid of light mode because everyone hates it)
ADDED COLOUR OPTIONS THEY LOOK SO COOL AHHH
Also depending on what computer you have, the faster the timer will be. NO that isn't a joke. It's because the code uses delta time and averages it out to create a more accurate timer. So actually it should be fine.

Also added support for the hour which wasn't working before.

NOTE: Do not use this as an actual speedrun timer unless you screen record it for the corner of a video or you are using it on a second monitor/different computer.
Also it's just a tad inaccurate sometimes so be aware of that. That's it for now :)))